+++ transport/roundtrip_test.go
@@ -72,12 +72,11 @@ func (r staticRegistry) ServiceProcedures() []transport.ServiceProcedure {
 	return []transport.ServiceProcedure{{Service: testService, Procedure: testProcedure}}
 }
 
-func (r staticRegistry) GetHandlerSpec(service string, procedure string) (transport.HandlerSpec, error) {
-	if procedure == testProcedure {
+func (r staticRegistry) Choose(ctx context.Context, req *transport.Request) (transport.HandlerSpec, error) {
+	if req.Procedure == testProcedure {
 		return transport.NewUnaryHandlerSpec(r.Handler), nil
-	} else {
-		return transport.NewOnewayHandlerSpec(r.OnewayHandler), nil
 	}
+	return transport.NewOnewayHandlerSpec(r.OnewayHandler), nil
 }
 
 // handlerFunc wraps a function into a transport.Registry